backend/requirements.txt
aenum==2.2.3
Flask==1.0.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
funkybob==2018.5.1
oauthlib==3.0.1
pyjwt==1.7.1
pymysql==0.9.3
pyOpenSSL==19.0.0
requests==2.21.0

# Some handy dev dependencies
ipdb==0.13.2
ipython==7.13.0
